% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggConfusionMatrix.R
\name{ggConfusionMatrix}
\alias{ggConfusionMatrix}
\title{Plot confusion matrix with metrics}
\usage{
ggConfusionMatrix(m)
}
\arguments{
\item{m}{An object of class confusionMatrix}
}
\value{
A \code{ggplot2} graphic object
}
\description{
Plot confusion matrix with metrics
}
