% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMeanDecreaseAccuracy.R
\name{ggMeanDecreaseAccuracy}
\alias{ggMeanDecreaseAccuracy}
\title{ggplot Mean Decrease Accuracy}
\usage{
ggMeanDecreaseAccuracy(model, num, ...)
}
\arguments{
\item{model}{An object of class \code{randomForest}}

\item{num}{An integer with the numbers of features to plot}

\item{...}{Other arguments}
}
\value{
A \code{ggplot2} graphic object with feature accuracy in
 decrease order, or a warning if model \code{importance} is set
 to \code{FALSE}
}
\description{
ggplot Mean Decrease Accuracy
}
\examples{
\dontrun{
# model is a randomForest model
ggMeanDecreaseAccuracy(model$finalModel)
}
}
