% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMeanDecreaseGini.R
\name{ggMeanDecreaseGini}
\alias{ggMeanDecreaseGini}
\title{ggplot Mean Decrease Gini}
\usage{
ggMeanDecreaseGini(model, num, ...)
}
\arguments{
\item{model}{An object of class \code{randomForest}}

\item{num}{An integer with the numbers of features to plot}

\item{...}{Other arguments}
}
\value{
A \code{ggplot2} graphic object, or a warning if model
 \code{importance} is set to \code{FALSE}
}
\description{
ggplot Mean Decrease Gini
}
\examples{
\dontrun{
# model is a randomForest model
ggMeanDecreaseGini(model$finalModel)
}
}
