% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotConfusionMatrix.R
\name{ggplotConfusionMatrix}
\alias{ggplotConfusionMatrix}
\title{Plot confusion matrix with metrics}
\usage{
ggplotConfusionMatrix(m)
}
\arguments{
\item{m}{An object of class confusionMatrix}
}
\value{
a plot
}
\description{
Plot confusion matrix with metrics
}
\examples{

}
