% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotModel.R
\name{ggplotModel}
\alias{ggplotModel}
\title{Plot model error and OOB}
\usage{
ggplotModel(model, ...)
}
\arguments{
\item{model}{An object of class `randomForest`}

\item{...}{Optional parameters to be passed}
}
\value{
A `ggplot2` graphic object
}
\description{
Plot model error and OOB
}
\examples{
\dontrun{
ggplotModel(model$finalModel)
}
}
