% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurface.r
\name{ggsurface}
\alias{ggsurface}
\title{Plot Model Surface with ggplot2}
\usage{
ggsurface(m, x.name, y.name, legend.title, col.low, col.high)
}
\arguments{
\item{m}{an glm or lm R base object model}

\item{x.name}{a name for the x axis. Default = NULL. The var name from the model will be used if no value is provided}

\item{y.name}{a name for the x axis. Default = NULL. The var name from the model will be used if no value is provided}

\item{legend.title}{a name for the scale surface values. Default = NULL. If no values is provided the response variable name from the model is used.}

\item{col.low}{the color of the response var lowest value. It might be a standard color name, hexadecimal or rgb generated (from the function rgb()).}

\item{col.high}{the color of the response var highest value. It might be a standard color name, hexadecimal or rgb generated (from the function rgb()).}
}
\description{
a ggplot2 surface plot function for R base glm or lm models
}
\examples{
m <- glm(mpg ~ wt + hp, data = mtcars)
ggsurface(m, x.name="Weight", y.name="Horsepower",
legend.title="Miles per Gallon", col.low = "white",
col.high = "blue")
}
\author{
Wilson Frantine-Silva
}
