% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotModel.R
\name{ggplotModel}
\alias{ggplotModel}
\title{Plot model error and OOB}
\usage{
ggplotModel(model, ...)
}
\arguments{
\item{model}{An object of class \code{randomForest}}

\item{...}{Optional parameters to be passed}
}
\value{
A \code{ggplot2} graphic object
}
\description{
Plot model error and OOB
}
\examples{
\dontrun{
# model is a randomForest model
ggplotModel(model$finalModel)
}
}
